chore: refactored getHumanToMs and parseValue function to utils
