+++ src/job-utils.js
@@ -1,3 +1,6 @@
+const ms = require('ms');
+const humanInterval = require('human-interval');
+const later = require('@breejs/later');
 const isSANB = require('is-string-and-not-blank');
 
 const isSchedule = (value) => {
@@ -10,5 +13,32 @@ const getName = (job) => {
   if (typeof job === 'function' && isSANB(job.name)) return job.name;
 };
 
+const getHumanToMs = (_value) => {
+  const value = humanInterval(_value);
+  if (Number.isNaN(value)) return ms(_value);
+  return value;
+};
+
+const parseValue = (value) => {
+  if (value === false) return value;
+
+  if (this.isSchedule(value)) return value;
+
+  if (isSANB(value)) {
+    const schedule = later.schedule(later.parse.text(value));
+    if (schedule.isValid()) return later.parse.text(value);
+    value = getHumanToMs(value);
+  }
+
+  if (!Number.isFinite(value) || value < 0)
+    throw new Error(
+      `Value ${value} must be a finite number >= 0 or a String parseable by \`later.parse.text\` (see <https://breejs.github.io/later/parsers.html#text> for examples)`
+    );
+
+  return value;
+};
+
 module.exports.isSchedule = isSchedule;
 module.exports.getName = getName;
+module.exports.getHumanToMs = getHumanToMs;
+module.exports.parseValue = parseValue;
